refactor: move perform and perform_closed_loop to base
