+++ src/aind_dynamic_foraging_models/generative_model/agent_q_learning.py
@@ -98,72 +98,6 @@ class ForagerSimpleQ(DynamicForagingAgentMLEBase):
         self.choice_kernel = np.full([self.n_actions, self.n_trials + 1], np.nan)
         self.choice_kernel[:, 0] = 0  # Initial choice kernel as 0
 
-    def perform(
-        self,
-        task: DynamicForagingTaskBase,
-    ):
-        """Generative simulation of a task, or "open-loop" simulation
-
-        Override the base class method to include choice_prob caching etc.
-        """
-        self.task = task
-        self.n_trials = task.num_trials
-
-        # --- Main task loop ---
-        self._reset()  # Reset agent
-        _, _ = self.task.reset()  # Reset task and get the initial observation
-        task_done = False
-        while not task_done:
-            assert self.trial == self.task.trial  # Ensure the two timers are in sync
-
-            # -- Agent performs an action
-            choice, choice_prob = self.act(_)
-
-            # -- Environment steps (enviromnet's timer ticks here!!!)
-            _, reward, task_done, _, _ = task.step(choice)
-
-            # -- Update agent history
-            self.choice_prob[:, self.trial] = choice_prob
-            self.choice_history[self.trial] = choice
-            # In Sutton & Barto's convention, reward belongs to the next time step, but we put it
-            # in the current time step for the sake of consistency with neuroscience convention
-            self.reward_history[self.trial] = reward
-
-            # -- Agent's timer ticks here !!!
-            self.trial += 1
-
-            # -- Update q values
-            # Note that this will update the q values **after the last trial**, a final update that
-            # will not be used to make the next action (because task is *done*) but may be used for
-            # correlating with physiology recordings
-            self.learn(_, choice, reward, _, task_done)
-
-    def perform_closed_loop(self, fit_choice_history, fit_reward_history):
-        """Simulates the agent over a fixed choice and reward history using its params.
-        Also called "teacher forcing" or "closed-loop" simulation.
-
-        Unlike .perform() ("generative" simulation), this is called "predictive" simulation,
-        which does not need a task and is used for model fitting.
-        """
-        self.n_trials = len(fit_choice_history)
-        self._reset()
-
-        while self.trial <= self.n_trials - 1:
-            # -- Compute and cache choice_prob (key to model fitting)
-            _, choice_prob = self.act(None)
-            self.choice_prob[:, self.trial] = choice_prob
-
-            # -- Clamp history to fit_history
-            clamped_choice = fit_choice_history[self.trial].astype(int)
-            clamped_reward = fit_reward_history[self.trial]
-            self.choice_history[self.trial] = clamped_choice
-            self.reward_history[self.trial] = clamped_reward
-
-            # -- Agent's timer ticks here !!!
-            self.trial += 1
-
-            # -- Update q values
-            self.learn(None, clamped_choice, clamped_reward, None, None)
 
     def act(self, _):
         """Action selection"""
